补充红包插件源码文档
